+++ src/main/java/gtPlusPlus/core/entity/monster/EntityBatKing.java
@@ -611,7 +611,7 @@ public class EntityBatKing extends EntityMob implements IRangedAttackMob {
             byte b0 = (byte) (this.attackCounter > 10 ? 1 : 0);
 
             if (b1 != b0) {
-                this.dataWatcher.updateObject(16, Byte.valueOf(b0));
+                this.dataWatcher.updateObject(16, b0);
             }
         }
     }
