+++ src/main/java/gtPlusPlus/core/entity/monster/EntityStaballoyConstruct.java
@@ -77,7 +77,7 @@ public class EntityStaballoyConstruct extends EntityIronGolem {
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(17, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(17, (byte) 0);
     }
 
     /**
